requirements.txt
requests
plotly_express==0.4.0